quick code reorg to make room for snippets too
